chore: remove produce block apis no longer part of spec (#7749)
**Motivation**

Closes https://github.com/ChainSafe/lodestar/issues/7410

**Description**

- removes `GET /eth/v2/validator/blocks/{slot}` and `GET
/eth/v1/validator/blinded_blocks/{slot}`
- deprecates `--useProduceBlockV3` flag, we will always use v3 now
