+++ packages/api/src/beacon/routes/validator.ts
@@ -291,35 +291,6 @@ export type Endpoints = {
    * Metadata in the response indicates the type of block produced, and the supported types of block
    * will be added to as forks progress.
    */
-  produceBlockV2: Endpoint<
-    "GET",
-    {
-      /** The slot for which the block should be proposed */
-      slot: Slot;
-      /** The validator's randao reveal value */
-      randaoReveal: BLSSignature;
-      /** Arbitrary data validator wants to include in block */
-      graffiti?: string;
-    } & Omit<ExtraProduceBlockOpts, "blindedLocal">,
-    {
-      params: {slot: number};
-      query: {
-        randao_reveal: string;
-        graffiti?: string;
-        fee_recipient?: string;
-        builder_selection?: string;
-        strict_fee_recipient_check?: boolean;
-      };
-    },
-    BeaconBlockOrContents,
-    VersionMeta
-  >;
-
-  /**
-   * Requests a beacon node to produce a valid block, which can then be signed by a validator.
-   * Metadata in the response indicates the type of block produced, and the supported types of block
-   * will be added to as forks progress.
-   */
   produceBlockV3: Endpoint<
     "GET",
     {
@@ -347,18 +318,6 @@ export type Endpoints = {
     },
     BeaconBlockOrContents | BlindedBeaconBlock,
     ProduceBlockV3Meta
-  >;
-
-  produceBlindedBlock: Endpoint<
-    "GET",
-    {
-      slot: Slot;
-      randaoReveal: BLSSignature;
-      graffiti?: string;
-    },
-    {params: {slot: number}; query: {randao_reveal: string; graffiti?: string}},
-    BlindedBeaconBlock,
-    VersionMeta
   >;
 
   /**
@@ -614,49 +573,6 @@ export function getDefinitions(config: ChainForkConfig): RouteDefinitions<Endpoi
         meta: ExecutionOptimisticCodec,
       },
     },
-    produceBlockV2: {
-      url: "/eth/v2/validator/blocks/{slot}",
-      method: "GET",
-      req: {
-        writeReq: ({slot, randaoReveal, graffiti, feeRecipient, builderSelection, strictFeeRecipientCheck}) => ({
-          params: {slot},
-          query: {
-            randao_reveal: toHex(randaoReveal),
-            graffiti: toGraffitiHex(graffiti),
-            fee_recipient: feeRecipient,
-            builder_selection: builderSelection,
-            strict_fee_recipient_check: strictFeeRecipientCheck,
-          },
-        }),
-        parseReq: ({params, query}) => ({
-          slot: params.slot,
-          randaoReveal: fromHex(query.randao_reveal),
-          graffiti: fromGraffitiHex(query.graffiti),
-          feeRecipient: query.fee_recipient,
-          builderSelection: query.builder_selection as BuilderSelection,
-          strictFeeRecipientCheck: query.strict_fee_recipient_check,
-        }),
-        schema: {
-          params: {slot: Schema.UintRequired},
-          query: {
-            randao_reveal: Schema.StringRequired,
-            graffiti: Schema.String,
-            fee_recipient: Schema.String,
-            builder_selection: Schema.String,
-            strict_fee_recipient_check: Schema.Boolean,
-          },
-        },
-      },
-      resp: {
-        data: WithVersion(
-          (fork) =>
-            (isForkPostDeneb(fork)
-              ? sszTypesFor(fork).BlockContents
-              : ssz[fork].BeaconBlock) as Type<BeaconBlockOrContents>
-        ),
-        meta: VersionCodec,
-      },
-    },
     produceBlockV3: {
       url: "/eth/v3/validator/blocks/{slot}",
       method: "GET",
@@ -757,32 +673,6 @@ export function getDefinitions(config: ChainForkConfig): RouteDefinitions<Endpoi
             };
           },
         },
-      },
-    },
-    produceBlindedBlock: {
-      url: "/eth/v1/validator/blinded_blocks/{slot}",
-      method: "GET",
-      req: {
-        writeReq: ({slot, randaoReveal, graffiti}) => ({
-          params: {slot},
-          query: {randao_reveal: toHex(randaoReveal), graffiti: toGraffitiHex(graffiti)},
-        }),
-        parseReq: ({params, query}) => ({
-          slot: params.slot,
-          randaoReveal: fromHex(query.randao_reveal),
-          graffiti: fromGraffitiHex(query.graffiti),
-        }),
-        schema: {
-          params: {slot: Schema.UintRequired},
-          query: {
-            randao_reveal: Schema.StringRequired,
-            graffiti: Schema.String,
-          },
-        },
-      },
-      resp: {
-        data: WithVersion((fork) => getPostBellatrixForkTypes(fork).BlindedBeaconBlock),
-        meta: VersionCodec,
       },
     },
     produceAttestationData: {
