+++ packages/beacon-node/src/api/impl/validator/index.ts
@@ -852,11 +852,6 @@ export function getValidatorApi(
   }
 
   return {
-    async produceBlockV2({slot, randaoReveal, graffiti, ...opts}) {
-      const {data, ...meta} = await produceEngineFullBlockOrContents(slot, randaoReveal, graffiti, opts);
-      return {data, meta};
-    },
-
     async produceBlockV3({slot, randaoReveal, graffiti, skipRandaoVerification, builderBoostFactor, ...opts}) {
       const {data, ...meta} = await produceEngineOrBuilderBlock(
         slot,
@@ -889,26 +884,6 @@ export function getValidatorApi(
       }
 
       return {data, meta};
-    },
-
-    async produceBlindedBlock({slot, randaoReveal, graffiti}) {
-      const {data, version} = await produceEngineOrBuilderBlock(slot, randaoReveal, graffiti);
-      if (!isForkPostBellatrix(version)) {
-        throw Error(`Invalid fork=${version} for produceBlindedBlock`);
-      }
-
-      if (isBlockContents(data)) {
-        const {block} = data;
-        const blindedBlock = beaconBlockToBlinded(config, block as BeaconBlock<ForkPostBellatrix>);
-        return {data: blindedBlock, meta: {version}};
-      }
-
-      if (isBlindedBeaconBlock(data)) {
-        return {data, meta: {version}};
-      }
-
-      const blindedBlock = beaconBlockToBlinded(config, data as BeaconBlock<ForkPostBellatrix>);
-      return {data: blindedBlock, meta: {version}};
     },
 
     async produceAttestationData({committeeIndex, slot}) {
