+++ packages/beacon-node/src/chain/chain.ts
@@ -166,9 +166,9 @@ export class BeaconChain implements IBeaconChain {
   /** Map keyed by executionPayload.blockHash of the block for those blobs */
   readonly producedContentsCache = new Map<BlockHash, deneb.Contents>();
 
-  // Cache payload from the local execution so that produceBlindedBlock or produceBlockV3 and
-  // send and get signed/published blinded versions which beacon can assemble into full before
-  // actual publish
+  // Cache payloads from the local execution so that we can send
+  // and get signed/published blinded versions which beacon node can
+  // assemble into full blocks before publishing to the network.
   readonly producedBlockRoot = new Map<RootHex, ExecutionPayload | null>();
   readonly producedBlindedBlockRoot = new Set<RootHex>();
   readonly blacklistedBlocks: Map<RootHex, Slot | null>;
