+++ packages/cli/src/cmds/validator/handler.ts
@@ -170,7 +170,6 @@ export async function validatorHandler(args: IValidatorCliArgs & GlobalArgs): Pr
       disableAttestationGrouping: args.disableAttestationGrouping,
       valProposerConfig,
       distributed: args.distributed,
-      useProduceBlockV3: args.useProduceBlockV3,
       broadcastValidation: parseBroadcastValidation(args.broadcastValidation),
       blindedLocal: args.blindedLocal,
       externalSigner: {
