+++ packages/validator/src/validator.ts
@@ -61,7 +61,6 @@ export type ValidatorOptions = {
   closed?: boolean;
   valProposerConfig?: ValidatorProposerConfig;
   distributed?: boolean;
-  useProduceBlockV3?: boolean;
   broadcastValidation?: routes.beacon.BroadcastValidation;
   blindedLocal?: boolean;
   externalSigner?: ExternalSignerOptions;
@@ -226,7 +225,6 @@ export class Validator {
     const syncingStatusTracker = new SyncingStatusTracker(logger, api, clock, metrics);
 
     const blockProposingService = new BlockProposingService(config, loggerVc, api, clock, validatorStore, metrics, {
-      useProduceBlockV3: opts.useProduceBlockV3,
       broadcastValidation: opts.broadcastValidation ?? defaultOptions.broadcastValidation,
       blindedLocal: opts.blindedLocal ?? defaultOptions.blindedLocal,
     });
@@ -314,15 +312,13 @@ export class Validator {
     await assertEqualGenesis(opts, genesis);
     logger.info("Verified connected beacon node and validator have the same genesisValidatorRoot");
 
-    const {useProduceBlockV3, broadcastValidation = defaultOptions.broadcastValidation, valProposerConfig} = opts;
+    const {broadcastValidation = defaultOptions.broadcastValidation, valProposerConfig} = opts;
     const defaultBuilderSelection =
       valProposerConfig?.defaultConfig.builder?.selection ?? defaultOptions.builderSelection;
     const strictFeeRecipientCheck = valProposerConfig?.defaultConfig.strictFeeRecipientCheck ?? false;
     const suggestedFeeRecipient = valProposerConfig?.defaultConfig.feeRecipient ?? defaultOptions.suggestedFeeRecipient;
 
     logger.info("Initializing validator", {
-      // if no explicit option is provided, useProduceBlockV3 will be auto enabled on/post deneb
-      useProduceBlockV3: useProduceBlockV3 === undefined ? "deneb+" : useProduceBlockV3,
       broadcastValidation,
       defaultBuilderSelection,
       suggestedFeeRecipient,
